chore(daemon): reorg default configs
